Change the plugin idea from airplane made to wp core blocker
