+++ CHANGES.md
@@ -1,3 +1,6 @@
+#### Version 0.2.0 - 2016/08/25
+* Rewrite plugin to only block core connections and disable plugin / theme installation
+
 #### Version 0.1.9 - 2016/07/25
 * Prevent BuddyPress from falling back to Gravatar. props @johnbillion
